requirements.txt
# Core Streamlit and data processing
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization libraries
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Databricks integrations
databricks-sql-connector>=3.0.0
databricks-sdk>=0.18.0

# Optional: Additional useful libraries
python-dotenv>=1.0.0
requests>=2.31.0
